third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/add-new-rule-with-style-after-body.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>
function addStyle()
{
    var style = document.createElement("style");
    document.documentElement.appendChild(style);
    style.sheet.insertRule("foo {display: none;}", 0);
}

function test() {
  TestRunner.cssModel.addEventListener(SDK.CSSModel.Events.StyleSheetAdded, stylesheetAdded);
  TestRunner.evaluateInPage('addStyle()');

  function stylesheetAdded() {
    TestRunner.cssModel.removeEventListener(SDK.CSSModel.Events.StyleSheetAdded, stylesheetAdded);
    ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step1);
  }

  var treeElement;
  var hasResourceChanged;

  function step1() {
    ElementsTestRunner.addNewRule('inspected', step2);
  }

  function step2() {
    var section = ElementsTestRunner.firstMatchedStyleSection();
    var newProperty = section.addNewBlankProperty();
    newProperty.startEditing();
    newProperty.nameElement.textContent = 'color';
    newProperty.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    newProperty.valueElement.textContent = 'maroon';
    newProperty.valueElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    ElementsTestRunner.waitForStyles('inspected', step3);
  }

  function step3() {
    TestRunner.addResult('After adding new rule:');
    ElementsTestRunner.dumpSelectedElementStyles(true, false, true);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a new rule works when there is a STYLE element after BODY. TIMEOUT SHOULD NOT OCCUR! <a href="https://bugs.webkit.org/show_bug.cgi?id=111299">Bug 111299</a>
</p>

<div id="inspected" style="font-size: 12px">Text</div>

</body>
</html>
