third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/add-new-rule-with-style-after-body-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that adding a new rule works when there is a STYLE element after BODY. TIMEOUT SHOULD NOT OCCUR! Bug 111299 https://bugs.webkit.org/show_bug.cgi?id=111299

After adding new rule:
[expanded] 
element.style { ()
    font-size: 12px;

[expanded] [no-affect] 
inspected { (inspector-stylesheet:1 -> inspector-stylesheet:1:12)
    color: maroon;
    : ;

[expanded] 
div { (user agent stylesheet)
    display: block;
